Move the __tests__ suite to the volto repo
