+++ packages/volto/__tests__/fixtures/test-volto-project/addons/my-volto-config-addon/package.json
