+++ packages/volto/__tests__/fixtures/test-volto-project/addons/non-volto-addon-lib/package.json
