+++ packages/volto/__tests__/fixtures/test-volto-project/addons/test-addon/package.json
