+++ packages/volto/__tests__/fixtures/test-volto-project/addons/test-addon/src/custom-addons/volto/server.jsx
