+++ packages/volto/__tests__/fixtures/test-volto-project/addons/test-released-dummy/index.js
