+++ packages/volto/__tests__/fixtures/test-volto-project/packages/test-local-packages-via-addons-env-var/package.json
