+++ packages/volto/__tests__/fixtures/test-volto-project/src/custom-addons/test-addon/testaddon.js
