+++ packages/volto/__tests__/fixtures/test-volto-project/src/custom-addons/volto/client.js
