+++ packages/volto/__tests__/fixtures/test-volto-project/src/customizations/server.jsx
