+++ packages/volto/__tests__/fixtures/test-volto-project/volto.config.js
