+++ packages/volto/__tests__/fixtures/volto-addon1.js
