+++ packages/volto/__tests__/fixtures/volto-addon2.js
