+++ packages/volto/__tests__/fixtures/volto-addon3.js
